requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
cssselect2==0.7.0
distro==1.9.0
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask_weasyprint==1.1.0
fonttools==4.53.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
openai==1.35.14
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.11.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
ruff==0.5.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tinycss2==1.3.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
zopfli==0.2.3
